requirements.txt
torch
transformers
Pillow
fastmcp
langchain
langchain-community
chromadb
typing
